refactor(swap): clean and simplify code
